identity/requirements/core.txt
Django==4.0.3
django-oauth-toolkit==1.7.0
djangorestframework==3.13.1
django-cors-headers==3.11.0
django-filter==21.1
